[!!!][TASK] Add custom node generator implementation
Node Generator are used to generate a specific node, based on the NodeType
each node type can have their own implementation. Now the Preset for content
and document node type must be an array. This allow the creation of node
with a random node type.

Breaking because, now the content and document node type must be an array
in your settings.yaml